+++ src/Caliburn.Light.Core/ViewModel/ViewAware.cs
@@ -9,24 +9,17 @@ namespace Caliburn.Light;
 /// </summary>
 public class ViewAware : ParentAware, IViewAware
 {
-    private const string DefaultContext = "__default__";
-
-    private List<KeyValuePair<string, WeakReference>>? _views;
+    private List<KeyValuePair<string?, WeakReference>>? _views;
 
     void IViewAware.AttachView(object view, string? context)
     {
         ArgumentNullException.ThrowIfNull(view);
 
-        if (context is null)
-            context = DefaultContext;
-
-        if (_views is null)
-            _views = new List<KeyValuePair<string, WeakReference>>(1);
-
         Trace.TraceInformation("Attaching view {0} to {1}.", view, this);
+        _views ??= new List<KeyValuePair<string?, WeakReference>>(1);
         var index = _views.FindIndex(p => string.Equals(p.Key, context, StringComparison.Ordinal));
         if (index < 0)
-            _views.Add(new KeyValuePair<string, WeakReference>(context, new WeakReference(view)));
+            _views.Add(new KeyValuePair<string?, WeakReference>(context, new WeakReference(view)));
         else
             _views[index].Value.Target = view;
 
@@ -36,9 +29,6 @@ public class ViewAware : ParentAware, IViewAware
     bool IViewAware.DetachView(object view, string? context)
     {
         ArgumentNullException.ThrowIfNull(view);
-
-        if (context is null)
-            context = DefaultContext;
 
         Trace.TraceInformation("Detaching view {0} from {1}.", view, this);
         var detached = _views?.RemoveAll(p => string.Equals(p.Key, context, StringComparison.Ordinal)) > 0;
@@ -53,7 +43,7 @@ public class ViewAware : ParentAware, IViewAware
     /// </summary>
     /// <param name="view">The view.</param>
     /// <param name="context">The context in which the view appears.</param>
-    protected virtual void OnViewAttached(object view, string context)
+    protected virtual void OnViewAttached(object view, string? context)
     {
     }
 
@@ -62,20 +52,17 @@ public class ViewAware : ParentAware, IViewAware
     /// </summary>
     /// <param name="view">The view.</param>
     /// <param name="context">The context in which the view appears.</param>
-    protected virtual void OnViewDetached(object view, string context)
+    protected virtual void OnViewDetached(object view, string? context)
     {
     }
 
     object? IViewAware.GetView(string? context)
     {
-        if (context is null)
-            context = DefaultContext;
-
         var entry = _views?.Find(p => string.Equals(p.Key, context, StringComparison.Ordinal)) ?? default;
         return entry.Value?.Target;
     }
 
-    IEnumerable<KeyValuePair<string, object>> IViewAware.GetViews()
+    IEnumerable<KeyValuePair<string?, object>> IViewAware.GetViews()
     {
         if (_views is null)
             yield break;
@@ -84,7 +71,7 @@ public class ViewAware : ParentAware, IViewAware
         {
             var view = entry.Value.Target;
             if (view is not null)
-                yield return new KeyValuePair<string, object>(entry.Key, view);
+                yield return new KeyValuePair<string?, object>(entry.Key, view);
         }
     }
 }